+++ packages/experimental/Switch/src/__tests__/Switch.test.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
-import { Switch } from '../Switch';
+import { Switch } from '../Switch.win32';
+import { checkRenderConsistency, checkReRender } from '@fluentui-react-native/test-tools';
 
 it('Switch Default', () => {
   const tree = renderer.create(<Switch label="Default Switch" />).toJSON();
@@ -10,4 +11,12 @@ it('Switch Default', () => {
 it('Switch Disabled', () => {
   const tree = renderer.create(<Switch label="Default Switch Disabled" disabled />).toJSON();
   expect(tree).toMatchSnapshot();
+});
+
+it('Switch simple rendering does not invalidate styling', () => {
+  checkRenderConsistency(() => <Switch label="Default Switch" />, 2);
+});
+
+it('Switch re-renders correctly', () => {
+  checkReRender(() => <Switch label="Render Twice" />, 2);
 });
